Fix missed test refactor.
