+++ kaldb/src/test/java/com/slack/kaldb/chunkManager/ChunkRollOverStrategyImplTest.java
@@ -1,4 +1,4 @@
-package com.slack.kaldb.chunk;
+package com.slack.kaldb.chunkManager;
 
 import static org.assertj.core.api.Assertions.assertThat;
